refactored SQL dumping script
