+++ .gitignore
@@ -1,3 +1,4 @@
 log
 
 .idea
+__pycache__/
